+++ consul/consul.go
@@ -1,6 +1,8 @@
 package consul
 
 import (
+	"errors"
+
 	"github.com/hashicorp/consul/api"
 
 	"github.com/philippgille/gokv/util"
@@ -8,15 +10,25 @@ import (
 
 // Client is a gokv.Store implementation for Consul.
 type Client struct {
-	c      *api.KV
-	folder string
+	c             *api.KV
+	folder        string
+	marshalFormat MarshalFormat
 }
 
 // Set stores the given value for the given key.
 // Values are marshalled to JSON automatically.
 func (c Client) Set(k string, v interface{}) error {
 	// First turn the passed object into something that Consul can handle
-	data, err := util.ToJSON(v)
+	var data []byte
+	var err error
+	switch c.marshalFormat {
+	case JSON:
+		data, err = util.ToJSON(v)
+	case Gob:
+		data, err = util.ToGob(v)
+	default:
+		return errors.New("The store seems to be configured with a marshal format that's not implemented yet")
+	}
 	if err != nil {
 		return err
 	}
@@ -54,7 +66,14 @@ func (c Client) Get(k string, v interface{}) (bool, error) {
 	}
 	data := kvPair.Value
 
-	return true, util.FromJSON(data, v)
+	switch c.marshalFormat {
+	case JSON:
+		return true, util.FromJSON(data, v)
+	case Gob:
+		return true, util.FromGob(data, v)
+	default:
+		return true, errors.New("The store seems to be configured with a marshal format that's not implemented yet")
+	}
 }
 
 // Delete deletes the stored value for the given key.
@@ -66,6 +85,16 @@ func (c Client) Delete(k string) error {
 	_, err := c.c.Delete(k, nil)
 	return err
 }
+
+// MarshalFormat is an enum for the available (un-)marshal formats of this gokv.Store implementation.
+type MarshalFormat int
+
+const (
+	// JSON is the MarshalFormat for (un-)marshalling to/from JSON
+	JSON MarshalFormat = iota
+	// Gob is the MarshalFormat for (un-)marshalling to/from gob
+	Gob
+)
 
 // Options are the options for the Consul client.
 type Options struct {
@@ -79,14 +108,18 @@ type Options struct {
 	// The Consul UI calls this "folder".
 	// Optional (none by default).
 	Folder string
+	// (Un-)marshal format.
+	// Optional (JSON by default).
+	MarshalFormat MarshalFormat
 }
 
 // DefaultOptions is an Options object with default values.
-// Scheme: "http", Address: "127.0.0.1:8500"
+// Scheme: "http", Address: "127.0.0.1:8500", Folder: none, MarshalFormat: JSON
 var DefaultOptions = Options{
 	Scheme:  "http",
 	Address: "127.0.0.1:8500",
-	// No need to define Folder, because its zero value is fine
+	// No need to define Folder because its zero value is fine
+	// No need to set MarshalFormat to JSON because its zero value is fine.
 }
 
 // NewClient creates a new Consul client.
@@ -110,8 +143,9 @@ func NewClient(options Options) (Client, error) {
 	}
 
 	result = Client{
-		c:      client.KV(),
-		folder: options.Folder,
+		c:             client.KV(),
+		folder:        options.Folder,
+		marshalFormat: options.MarshalFormat,
 	}
 
 	return result, nil