+++ gomap/map.go
@@ -1,6 +1,7 @@
 package gomap
 
 import (
+	"errors"
 	"sync"
 
 	"github.com/philippgille/gokv/util"
@@ -8,17 +9,28 @@ import (
 
 // Store is a gokv.Store implementation for a Go map with a sync.RWMutex for concurrent access.
 type Store struct {
-	m    map[string][]byte
-	lock *sync.RWMutex
+	m             map[string][]byte
+	lock          *sync.RWMutex
+	marshalFormat MarshalFormat
 }
 
 // Set stores the given object for the given key.
 // Values are marshalled to JSON automatically.
 func (m Store) Set(k string, v interface{}) error {
-	data, err := util.ToJSON(v)
+	var data []byte
+	var err error
+	switch m.marshalFormat {
+	case JSON:
+		data, err = util.ToJSON(v)
+	case Gob:
+		data, err = util.ToGob(v)
+	default:
+		return errors.New("The store seems to be configured with a marshal format that's not implemented yet")
+	}
 	if err != nil {
 		return err
 	}
+
 	m.lock.Lock()
 	defer m.lock.Unlock()
 	m.m[k] = data
@@ -40,7 +52,14 @@ func (m Store) Get(k string, v interface{}) (bool, error) {
 		return false, nil
 	}
 
-	return true, util.FromJSON(data, v)
+	switch m.marshalFormat {
+	case JSON:
+		return true, util.FromJSON(data, v)
+	case Gob:
+		return true, util.FromGob(data, v)
+	default:
+		return true, errors.New("The store seems to be configured with a marshal format that's not implemented yet")
+	}
 }
 
 // Delete deletes the stored value for the given key.
@@ -50,10 +69,35 @@ func (m Store) Delete(k string) error {
 	return nil
 }
 
-// NewStore creates a new Go sync.Map store.
-func NewStore() Store {
+// MarshalFormat is an enum for the available (un-)marshal formats of this gokv.Store implementation.
+type MarshalFormat int
+
+const (
+	// JSON is the MarshalFormat for (un-)marshalling to/from JSON
+	JSON MarshalFormat = iota
+	// Gob is the MarshalFormat for (un-)marshalling to/from gob
+	Gob
+)
+
+// Options are the options for the Go map store.
+type Options struct {
+	// (Un-)marshal format.
+	// Optional (JSON by default).
+	MarshalFormat MarshalFormat
+}
+
+// DefaultOptions is an Options object with default values.
+// MarshalFormat: JSON
+var DefaultOptions = Options{
+	// No need to set MarshalFormat to JSON
+	// because its zero value is fine.
+}
+
+// NewStore creates a new Go map store.
+func NewStore(options Options) Store {
 	return Store{
-		m:    make(map[string][]byte),
-		lock: new(sync.RWMutex),
+		m:             make(map[string][]byte),
+		lock:          new(sync.RWMutex),
+		marshalFormat: options.MarshalFormat,
 	}
 }