+++ gomap/map_test.go
@@ -12,20 +12,38 @@ import (
 // TestStore tests if reading from, writing to and deleting from the store works properly.
 // A struct is used as value. See TestTypes() for a test that is simpler but tests all types.
 func TestStore(t *testing.T) {
-	store := gomap.NewStore()
-	test.TestStore(store, t)
+	// Test with JSON
+	t.Run("JSON", func(t *testing.T) {
+		store := createStore(t, gomap.JSON)
+		test.TestStore(store, t)
+	})
+
+	// Test with gob
+	t.Run("gob", func(t *testing.T) {
+		store := createStore(t, gomap.Gob)
+		test.TestStore(store, t)
+	})
 }
 
 // TestTypes tests if setting and getting values works with all Go types.
 func TestTypes(t *testing.T) {
-	store := gomap.NewStore()
-	test.TestTypes(store, t)
+	// Test with JSON
+	t.Run("JSON", func(t *testing.T) {
+		store := createStore(t, gomap.JSON)
+		test.TestTypes(store, t)
+	})
+
+	// Test with gob
+	t.Run("gob", func(t *testing.T) {
+		store := createStore(t, gomap.Gob)
+		test.TestTypes(store, t)
+	})
 }
 
 // TestStoreConcurrent launches a bunch of goroutines that concurrently work with one store.
 // The store is Go map with manual locking via sync.RWMutex, so testing this is important.
 func TestStoreConcurrent(t *testing.T) {
-	store := gomap.NewStore()
+	store := createStore(t, gomap.JSON)
 
 	goroutineCount := 1000
 
@@ -52,4 +70,29 @@ func TestStoreConcurrent(t *testing.T) {
 			t.Errorf("Expected: %v, but was: %v", expected, actual)
 		}
 	}
+}
+
+// TestErrors tests some error cases.
+func TestErrors(t *testing.T) {
+	// Test with a bad MarshalFormat enum value
+
+	store := createStore(t, gomap.MarshalFormat(19))
+	err := store.Set("foo", "bar")
+	if err == nil {
+		t.Error("An error should have occurred, but didn't")
+	}
+	// TODO: store some value for "foo", so retrieving the value works.
+	// Just the unmarshalling should fail.
+	// _, err = store.Get("foo", new(string))
+	// if err == nil {
+	// 	t.Error("An error should have occurred, but didn't")
+	// }
+}
+
+func createStore(t *testing.T, mf gomap.MarshalFormat) gomap.Store {
+	options := gomap.Options{
+		MarshalFormat: mf,
+	}
+	store := gomap.NewStore(options)
+	return store
 }