+++ syncmap/map_test.go
@@ -11,20 +11,38 @@ import (
 
 // TestStore tests if reading and writing to the store works properly.
 func TestStore(t *testing.T) {
-	store := syncmap.NewStore()
-	test.TestStore(store, t)
+	// Test with JSON
+	t.Run("JSON", func(t *testing.T) {
+		store := createStore(t, syncmap.JSON)
+		test.TestStore(store, t)
+	})
+
+	// Test with gob
+	t.Run("gob", func(t *testing.T) {
+		store := createStore(t, syncmap.Gob)
+		test.TestStore(store, t)
+	})
 }
 
 // TestTypes tests if setting and getting values works with all Go types.
 func TestTypes(t *testing.T) {
-	store := syncmap.NewStore()
-	test.TestTypes(store, t)
+	// Test with JSON
+	t.Run("JSON", func(t *testing.T) {
+		store := createStore(t, syncmap.JSON)
+		test.TestTypes(store, t)
+	})
+
+	// Test with gob
+	t.Run("gob", func(t *testing.T) {
+		store := createStore(t, syncmap.Gob)
+		test.TestTypes(store, t)
+	})
 }
 
 // TestStoreConcurrent launches a bunch of goroutines that concurrently work with one store.
 // The store is a sync.Map, so the concurrency should be supported by the used package.
 func TestStoreConcurrent(t *testing.T) {
-	store := syncmap.NewStore()
+	store := createStore(t, syncmap.JSON)
 
 	goroutineCount := 1000
 
@@ -51,4 +69,29 @@ func TestStoreConcurrent(t *testing.T) {
 			t.Errorf("Expected: %v, but was: %v", expected, actual)
 		}
 	}
+}
+
+// TestErrors tests some error cases.
+func TestErrors(t *testing.T) {
+	// Test with a bad MarshalFormat enum value
+
+	store := createStore(t, syncmap.MarshalFormat(19))
+	err := store.Set("foo", "bar")
+	if err == nil {
+		t.Error("An error should have occurred, but didn't")
+	}
+	// TODO: store some value for "foo", so retrieving the value works.
+	// Just the unmarshalling should fail.
+	// _, err = store.Get("foo", new(string))
+	// if err == nil {
+	// 	t.Error("An error should have occurred, but didn't")
+	// }
+}
+
+func createStore(t *testing.T, mf syncmap.MarshalFormat) syncmap.Store {
+	options := syncmap.Options{
+		MarshalFormat: mf,
+	}
+	store := syncmap.NewStore(options)
+	return store
 }